Websocket chat app works! Finally `mix phoenix.new`'d so lots of files with this one.
The User code from before has been moved into UserService. It's currently being run as
a worker for the main application, which is a Phoenix application. I think there's a
good bit that can be skimmed out of this still, but I'm still learning the ropes.

Current status: returning a custom template that's populating a user name from the
UserService (awesome). Even more awesome: includes a simple JS client that connects
to server using websockets.